requirements.txt
python-dotenv==0.15.0
celery==4.4.7
flask==1.1.2
redis==3.5.3
psycopg2-binary==2.8.6
SQLAlchemy==1.3.23
Flask==1.1.2
Faker==6.6.0
python-dotenv==0.15.0
Flask-UUID==0.2
flower==0.9.7